Move stable10 encryption steps to OccContext
